refactor: move Genshin UI menu to UI folder
