+++ setup_wizard/ui/ui_setup_wizard_menu.py
